+++ src/main/java/webserver/CustomHandler/LoginHandler.java
@@ -1,12 +1,11 @@
 package webserver.CustomHandler;
 
 import db.MemoryUserRepository;
+import http.request.HttpRequest;
+import http.response.HttpResponse;
 import model.User;
-import webserver.HttpRequest;
-import webserver.HttpResponse;
 
 import java.io.IOException;
-import java.util.Map;
 
 import static webserver.constant.Http.*;
